Separate module for sparse class and operations
