+++ heat/core/factories.py
@@ -17,8 +17,6 @@ from .types import datatype
 from . import devices
 from . import types
 
-from .Dcsr_matrix import Dcsr_matrix
-from scipy.sparse import csr_matrix as scipy_csr
 
 __all__ = [
     "arange",
@@ -36,111 +34,7 @@ __all__ = [
     "ones_like",
     "zeros",
     "zeros_like",
-    "sparse_csr_matrix",
 ]
-
-
-def sparse_csr_matrix(
-    obj: torch.sparse_csr_tensor,
-    dtype: Optional[Type[datatype]] = None,
-    copy: bool = True,
-    ndmin: int = 0,
-    order: str = "C",
-    split: Optional[int] = None,
-    is_split: Optional[int] = None,
-    device: Optional[Device] = None,
-    comm: Optional[Communication] = None,
-) -> Dcsr_matrix:
-
-    # TODO:
-    # Things I have not really paid attention to:
-    #   1. copy
-    #   2. ndim
-    #   3. order
-    #   4. balanced
-
-    # Convert input into torch.sparse_csr_tensor
-    if isinstance(obj, scipy_csr):
-        obj = torch.sparse_csr_tensor(obj.indptr, obj.indices, obj.data)
-
-    # For now, assuming the obj is torch.sparse_csr_tensor
-    comm = sanitize_comm(comm)
-    gshape = obj.shape
-    gnnz = obj.values().shape[0]
-
-    if split == 0:
-        start, end = comm.chunk(gshape, split, type="sparse")
-
-        # Find the starting and ending indices for
-        # col_indices and values tensors for this process
-        indicesStart = obj.crow_indices()[start]
-        indicesEnd = obj.crow_indices()[end]
-
-        # Slice the data belonging to this process
-        data = array(
-            obj=obj.values()[indicesStart:indicesEnd],
-            dtype=dtype,
-            copy=False,
-            split=None,
-            comm=comm,
-        )
-        # start:(end + 1) because indptr is of size (n + 1) for array with n rows
-        indptr = array(
-            obj=obj.crow_indices()[start : end + 1], dtype=dtype, copy=False, split=None, comm=comm
-        )
-        indices = array(
-            obj=obj.col_indices()[indicesStart:indicesEnd],
-            dtype=dtype,
-            copy=False,
-            split=None,
-            comm=comm,
-        )
-        lnnz = data.gshape[0]
-        indptr = indptr - indptr[0]
-
-    elif split is not None:
-        raise NotImplementedError("Not implemented for other splitting-axes")
-
-    elif is_split == 0:
-        data = array(
-            obj=obj.values(),
-            dtype=dtype,
-            copy=False,
-            split=None,
-            comm=comm,
-        )
-        indptr = array(obj=obj.crow_indices(), dtype=dtype, copy=False, split=None, comm=comm)
-        indices = array(
-            obj=obj.col_indices(),
-            dtype=dtype,
-            copy=False,
-            split=None,
-            comm=comm,
-        )
-        lnnz = data.gshape[0]
-
-    elif is_split is not None:
-        raise NotImplementedError("Not implemented for other splitting-axes")
-
-    else:  # split is None and is_split is None
-        data = array(obj=obj.values(), dtype=dtype, copy=False, split=None, comm=comm)
-        indptr = array(obj=obj.crow_indices(), dtype=dtype, copy=False, split=None, comm=comm)
-        indices = array(obj=obj.col_indices(), dtype=dtype, copy=False, split=None, comm=comm)
-        lnnz = gnnz
-
-    return Dcsr_matrix(
-        data=data,
-        indptr=indptr,
-        indices=indices,
-        gnnz=gnnz,
-        lnnz=lnnz,
-        gshape=gshape,
-        dtype=dtype,
-        split=split,
-        device=device,
-        comm=comm,
-        balanced=True,
-    )
 
 
 def arange(
