+++ heat/sparse/__init__.py
@@ -1,0 +1,6 @@
+"""
+add sparse heat function to the ht.sparse namespace
+"""
+
+from .Dcsr_matrix import *
+from .factories import *
